Signature provider moved to base
