+++ plugins/base/src/main/kotlin/signatures/SignatureProvider.kt
@@ -1,5 +1,6 @@
-package org.jetbrains.dokka.model
+package org.jetbrains.dokka.base.signatures
 
+import org.jetbrains.dokka.model.Documentable
 import org.jetbrains.dokka.pages.ContentNode
 
 interface SignatureProvider {
